Update tile generation with randomization
Previously, this function was applying the color classes in a linear manner, so the same tiles were on the same spot every time.
I solved this issue by creating a new array that holds twice the amount of colors, shuffling it and then using its content to assign the classes, without altering the original array in any way.