Implement transitions for improved tile click interactions
In this commit I aim to improve the user interface by adding a transition that takes places during the three scenarios of the game: tile reveal, tile matching and tile hiding. Additionally, I've slightly increased the setTimeout's duration so the transitions would last long enough for the user to see the tiles properly.